Added upload service for receiving object/scene files and storing them in S3.
Added hooks for restricting operations to the requesting user and
automatically setting userId on creation. Tested with Objects and
Scenes.

Changed association between Scenes and Objects to Many-to-Many.
Had to make a new service for xr_objects_scenes, the new table that
holds those many-to-many relationships.
Haven't yet added anything to automatically populate that relationship.

Fixed a mistyping of Locations' maxUsersPerInstance field.