Added public video service.
/public-video accepts single and arrays of JSON to store references
to videos.
On insertion, it will then take the Youtube URL it received and
download it in 1080p .webm format at < 3mbs, then upload that to
S3. When that's done, the SQL entry is updated so that the link is
pointing to the Cloudfront link to that video (currently the S3 link
since we're waiting to cache Cloudfront, simple to change).

Added node-config for more easily pulling nested options out of config
files.
Added aws section to config.

Added some config file names to .gitignore.