Added UUID generation hook. Updated how remote resource paths are generated to prevent collisions.
Upload service now generates UUID and uses it as an identifier for all
related resources. So, if a file is uploaded with a thumbnail, the
static-resource for the file will have that UUID, the stored path will
start with the UUID, and both the file and its thumbnail will be stored
under that UUID path, e.g. abcd1234/video/cool-video.mp4 and
abcd1234/image/cool-video-thumbnail.jpg

/video endpoint is still generating "public" videos, so they are all
grouped under the "UUID" of "public". Under that, it's now <service>/
<service-id>/<type>/<name>, e.g.
public/youtube/abcd1234_m/video/manifest.mpd.
This may change in the future once we implement a better way to indicate
public vs. private videos.