Fixed several incorrect references to identityProvider.type. It is now identityProvider.identityProviderType.
Removed hashing of identityProvider.token on creation. This broke a lot
of functionality that required looking up an identityProvider by token,
as well as password login's post-signup hook to send out the verify
email. There are probably some workarounds, but for now it's easier to
keep that as plaintext. The password, if it's present, is still hashed
via bcrypt.

Moved identityProvider hashing hook to separate file.