fixed movement
position was being updated as the same value as
velocity since they both referenced the same array.
Now a new array is created from the identity array instead.

The updatePosition behavior is now used to update
position from velocity through a subscription.

The IDLE and MOVEMENT states are now correctly updated
when the movement inputs are turned on/off.

The decelerate behavior was fixed and is called
properly since the MOVEMENT state is correct.