Started altering Mediasoup setup to enable multiple concurrent party video calls.
Existing setup was creating one router that handled all transports
and producers, meaning everyone on an instance server would be on
the same call, which is not what we want.

New setup has a router to handle instance-wide networking, such as
data transfer for instance state and possibly instance-wide chat
for showrooms, as well as on-demand creation of routes for parties.

It's sort of working, but switching instances is causing problems
and instance chat hasn't been tested yet.


Former-commit-id: 638a26f96b66b07e375cf8bef7943452e2a86ed3