Updated auto-connection logic, switched deployment target.
Problems with auto-connection may have stemmed from clients trying to
connect before their receive transport was ready. Sending of initial
producers is not part of the transport create handler anymore, but is
triggered by reception of new message 'WebRTCRequestCurrentProducers'.
On the client, this is sent when the receive transport changes to state
'connected'.

Added some conditional checks to PartyParticipantWindow on truncateUsername.
There are some times in local where the user might be undefined, and this was
causing uncaught errors.

Switched Github Actions to point to new TheOverlay deployment.


Former-commit-id: 6bedae144cd8b5e6e02c402e518f4ea90374506b