Worked around an an issue with swapping identity providers.
Previous behavior was that, upon logging into an OAuth account that
already had an identity provider (IP), the guest user/IP would be removed,
and the user would be logged in as the existing user from a JWT generated
from the existing IP.

For unclear reasons, something is calling JWT.authenticate after the server
redirects the user to the page they were on, but with the old JWT, which now
points to a nonexistent IP. Whatever is making this call does not properly
handle errors, and was crashing the server.

Workaround is to delete the guest user, but to verify the newer IP and remove
the old one. This makes it so that the 'guest' user's JWT still points to
a valid IP.

This might cause issues with multiple logins logging out the older session, as
that JWT is now pointed to a nonexistent IP, but a client auth call in this
situation should not crash anything.


Former-commit-id: be90ac76829ef952a5a101d6179679b3c6a29c3d