Re-enabled positional audio and fixed bugs in Harmony.
Location audio was not working because Positional audio system was not being started.
Uncommented lines in initialize.ts to re-enable it.

Made some minor fixes to loadAvatarForUpdatedUser. Needed to return Promise.resolve
if user not in an instance, and needed to check that both instanceId and channelInstanceId
are null in order to do that. Also made user.on('patched') only call it if the Network is
initialized.

Moved initialization of ClientNetworkStateSystem to be next to ClientNetworkSystem.
It was in the block that only gets intialized when the canvas is in use, and it can
be needed when there is not canvas (i.e. Harmony).

Fixed some bugs in Harmony:
* Needs to fire SCENE_LOADED manually to start Engine
* Listens to userState and triggers re-fetching layer users.
* provisionChannelServer listens for TRIGGER_UPDATE_CONSUMERS and calls triggerUpdateConsumers

Added a utility function waitForClientAuthenticated. This just loops until the client has been
authenticated and returns a promise. The Channel find call was occurring before the client had
credentials and was failing. If any other calls are failing for the same reason, this can be
await'ed before making the call.