Updated ServerTransport to use more than one worker/router.
Mediasoup can only handle about 500 consumers per worker. We were only creating a single worker,
so only using a single core. Updated SocketWebRTCServerTransport.ts to hold multiple workers.
WebRTCFunction:startWebRTC() now creates a worker per core on the host machine, and an instance
router per worker. Any time a transport is being made, it's made on the router with the fewest
transports. Producers are piped from the router they're on to all of the other routers so that
they can be consumed on any router.

Fixed an issue in new engine/src/initializeEngine.ts that was skipping registerClientSystems if
options.renderer.disable was true.

Made some selfUser.<field> checks conditional on selfUser.