Fixes for deployment process.
build-docker now prunes old containers and images.
Seeing if removing the `COPY . .` from Dockerfile-build-deploy
still works. The starting workdir of the builder image is /app, and
if we're putting the final image in /app, it should be the same directory,
meaning we don't have to copy anything.

Setting builder probe periodSeconds to 10 minutes, since sometimes the build process
seems to make the pod semi-unresponsive.