Updated api server cluster role with new permissions.
The API server needs to be able to patch the builder deployment or pods
when a project is updated so that it'll be rebuilt. Added create, update,
patch, and delete as allowed verbs, and added deployments as a core resource
that api-server can access.

Also added deployments and patch to gameserver cluster role just in case.