Improved Github actions and standardized scripts.
Made some changes so that forks can work without having to have
a commit at the head to change which/how GH actions run:

* Made all Github actions other than dev/prod-deploy disabled unless an
associated variable is set as a secret.

* Standardized order of PRIVATE_ECR and AWS_REGION between publish_ecr[_builder].sh
and build_docker[_builder].sh. Previous swapped order was confusing.

* Made deploy scripts and run-builder.sh always call build_docker[_builder].sh and
publish_ecr[_builder].sh with variables $PRIVATE_ECR and $AWS_REGION. If the fork is
using a private ECR repo, they should set this variable as a secret for the action to 'true',
and have the same variable set to 'true' in the builder Helm config. If the fork is using a
public ECR repo, then this doesn't need to be set at all in either place.

* Fixed incorrect usage of $AWS_REGION in publis_ecr[_builder].sh.