Made matchmaker create new location instance instead of new location.
Improved instanceProvision to handle if a GS is already assigned to an instance
but doesn't have any users yet, e.g. if a matchmaker has reserved it but users
haven't been placed on it.

Added instance-authorized-user service to handle instances that should only be available
for certain users, e.g. if it's a match server. Haven't tested it yet or integrated it
with match-ticket-assignment.

Updated GS channel connection logic to check if there's an instance already assigned to it.