Added security for identity-provider service.
GET/PATCH/UPDATE/REMOVE from external sources call a hook that ensures
the item being acted on is owned by the caller, and throws a 404 if it
is not. If it's a multi request, it ensures that the query has the userId
of the caller so they can't change anyone else's identity-providers.
FIND also ensures that the query has the userId of the caller, so they
only get their identity-providers.

Update CREATE service to disallow external creation of anything but
'guest' and 'password' identity-providers.