Added retry to S3 storage provider for NoSuchBucket errors.
Some storageProvider tests have been intermittently failing due to NoSuchBucket
errors from S3. These seem to be entirely due to records of a new bucket's existence not
propagating everywhere instantaneously.

Added a new method to s3.storage.ts, which performs a headBucket call (which seems to be more reliable
than when some other S3 method is called) and, if that doesn't return an error, calls waitFor('bucketExists').
When this resolves, it resolves the returned promise.

Most calls now use this in their error handling if they get a NoSuchBucket error. If it resolves successfully,
those functions re-call themselves.