Fixed some problems with storageProviders and their tests. (#4184)
local.storage.ts was returning an incorrect value for url from getSignedUrl; it had the key appended to the domain,
when S3 just returns the domain there. This url value was being used by listFolderContent to get the url of a file.
Now url is just the cacheDomain, and listFolderContent appends fields.Key to the url.

Storage provider tests have been improved. A configured bucket name is used for S3 tests. If the bucket doesn't exist,
it's created; otherwise it's left alone, as opposed to deleting and recreating a bucket for every test. S3 tests are
only run if STORAGE_S3_TEST_RESOURCE_BUCKET is non-null.

A UUID is part of the testFolderName so that multiple tests can write concurrently to the bucket without conflicting.
s3StorageAfterTest just wipes every file in the test folder, but leaves everything outside the folder alone, and does
not delete the bucket.

glb test runs inside the specified test folder so that it doesn't conflict with other runs of that test.

Removed duplication of testFolderName and folderKeyTemp's, instead passing them as parameter to some storageproviderconfig.ts
function calls.