Fixed some issues with party fetching.
User's party should be fetched with party.get(), not party.find().
party.find() changed to be restricted to admins.

party state not updating when result of party.get() is merged in. Not sure why.

Simplified party/index.tsx so that its returned DOM elements are rendered off of
a singular party, not an array of parties. Users are only ever in one party at a time.

Made selectedGroup.groupUsers copy itself with .map() before being sorted so that .sort is
not acting directly on hookState, which throws errors.