Did some touch-ups to the Show User ID selector.
Made the show/hide User ID selector not a button - it was kind of large and looked out-of-place.
Changed state/setState for userId state to userIdState/setUserIdState, which is more descriptive.
Added some styling to showUserId selector - pointer:cursor and hover color change.

Because it was bothering me, made styling for Logout selector that has color change on hover.