Improved instance provisioning to reduce double-provisions and other issues. (#4221)
Encountered situations where multiple instance provisions were being made at almost the same
time, resulting in a client trying to connect to the same gameserver for both instance and channel,
which caused nasty bugs.

Added two new columns to the instance table, 'assigned' and 'assignedAt', to be used in provisioning
free gameservers.

Getting a free gameserver now works like this:
* getFreeGameserver has a count, `iteration`, of how many times it has run.
* The IP address of a seemingly free gameserver is acquired in much the same way it was before, though
IPs already associated with an Assigned instance record are filtered out.
* An instance is made associated with that seemingly free IP address, with assigned=true and assignedAt set
to the current datetime.
* When that instance has been made, all non-ended instances associated with that IP address that have assigned=true are fetched.
* If there is more than one result, then only the earliest one to be assigned is the one that should actually be associated with
that gameserver.
* assignedAt values are compared; if the instance provision is not the earliest one to claim that gameserver, then that assigned
instance is deleted. If it is the earliest one, then it returns the IP address, and the provision can finish and return a valid
IP address to the client.
* An instance provision that was not first runs getFreeGameserver again, with iteration increased by 1.
* If an instance provision makes 10 attempts without being able to get a free gameserver, then it's assumed there are currently no
available gameservers, so it returns null (the client will handle this by waiting 10 seconds before trying to provision another instance).

There have also been a few calls inserted to delete any assigned instances older than 30 seconds - if a provision hasn't been connected to
in that long, it's assumed it never will be.

Gameservers now check to see if there's already an instance assigned to their IP address when the first connection is made. If not, then they
make a new instance as before (this should not happen now, but it's a fine enough fallback). If not, then they patch that assigned instance
with the locationId or channelId, and provision a subdomain for themself.

Fixed a bug with setting all local instances to 'ended' (and now 'assigned:false') on ServerTransport startup.

Fixed a client-side bug with provisioning an instance server after a prior attempt failed, and the client waited 10 seconds.