Fixed issues with audio/video not transmitting and local channel connection.
New audio and video was not being sent to clients because of an improper overwrite of variable 'userId'
in handleWebRtcSendTrack().

Fixed a bug where local channel servers restarted on a second user joining since null was being !== to
undefined, which returns true. Changed that to !=. We're just checking if two different values for locationId
are different, if they're both undefined/null then no need to restart.