Fixed issues with provisioning channel instances and clearing instanceIds.
Much of the front-end was relying on the user's channelInstanceId being correct for getting the instanceChannel
and provisioning an instance for it. However, users often had old channelInstanceIds on them, leading to provisioning
of channel instances that were incorrect. Checks being done to find the instance channel in a user's channel store
were only checking that there was a channel with channelType 'instance', which was often their old instance channel.

Updated instance provisioning to return instance ID as well as address. This ID is used for a number of checks, like
finding the channel that has that instanceId.

Fixed LOADED_CHANNEL handler, which was not returning the right values in order to find duplicates.

Fixed logic to patch user's (channel)instanceId when they leave an instance. It was relying on them still being in
Engine.defaultWorld.clients, but by the time they disconnect, they may no longer be there. It now just tries to
patch a user with that user's ID and that (channel)instanceId, which should ignore any other users and ignore that
user if they're already on a different instance.

Better handling of when to call connectToInstanceServer - prior code left possibility of trying to connect twice,
which caused bugs.

Removed some bot text from joined/left messages.