Fixed some bugs in user service and MediaStreamService.
Inventory system added post-find/get hooks for user service that parses stringified
inventory JSON. This wasn't taking into account if there were no inventory items to parse,
which has been corrected by checking if <data>.inventory_items exists before getting length.
Error handler for this was setting returned data to an empty array, which was not right.
It should just not do anything other than log the error, so that whatever user data was
being returned is returned as it was.

Further reflection on this issue showed that the reason it was occurring was that some of our
custom user find handlers were ignoring all params on the initial call, such as including
associated models. All user.find custom handlers were modified to just alter the query as needed
to get the specified user subset, such as doing a FindAndCountAll if a more complex SQL query
is needed and then adding `id: { $in: <returned ids> }` to the query.

user.GET hook was also written assuming the return structure of a FIND, which was incorrect.
GET just returns the single object that was found, or null

In MediaStreamService.ts, changed consumers.consumers to just consumers. There was no purpose
to this nested structure, and the one useEffect that was listening to MediaStreamService.consumers
wasn't updating properly as a result.