Update incoming / outgoing network tests
Due to decoupling between incoming and outgoing actions, there is much less logic to test. As far as actions are concerns, the outgoing system is dead simple. The incoming system is also much simpler, with the only addition being a caching mechanism. Rerouting of actions is no longer handled in these systems; instead, the gameserver simply adds actions to both incoming and outgoing queues as soon as it receives the messages.