Network Action Caching & Avatar Detail Refactor (#4394)
* start to refactor avatar loading stuff

* refactor works, just need to fix edge cases

* clean up

* action fix, avatar loading stuff, fix double register of golf system

* Add dispatchFromHost helper

* add joined world

* implement cached actions

* Implement client-authoritative network ids

* Decouple Incoming & Outgoing actions

* Update incoming / outgoing network tests

Due to decoupling between incoming and outgoing actions, there is much less logic to test. As far as actions are concerns, the outgoing system is dead simple. The incoming system is also much simpler, with the only addition being a caching mechanism. Rerouting of actions is no longer handled in these systems; instead, the gameserver simply adds actions to both incoming and outgoing queues as soon as it receives the messages.

* merge

* fix network integration tests

Co-authored-by: Gheric Speiginer <gheric.speiginer@gmail.com>