Fixed issues with client config.
VITE_* ENV_VARS need to be present in Docker build context in order for them to be built into the client.
The simplest solution was to write a new script, create-env-production.ts, that copies all VITE_* vars
into .env.production in the project root. Docker will copy this into the build context, and vite will
use all of the VITE_* vars when it builds.

Updated server and gameserver addresses in client to only use VITE_(GAME)SERVER_HOST in production, since
the port is only used for local development.