Updated xrengine-matchmaking Helm chart to work. (#4547)
* Updated xrengine-matchmaking Helm chart to work.

director and matchmaking code was updated to create a liveness/readiness
webserver so that the Helm deployment can properly monitor them. They
also source some variables, like the matchmaking function port, from
environment variables that the Helm deployment sets.

Made a Helm chart for the Open Match deployment that includes an ingress
for the frontend service, so that kubectl port-forward isn't necessary, and
so that a production deployment has an ingress for that traffic.

xrengine-matchmaking Helm chart updated to work. Removed frontend templates
since they're not needed. Put all of the matchmaking resources in their own
namespace.

* readme fix, chart path to xrengine-matchmaker

* Renamed xrengine-matchmaker to xrengine-matchmaking.

Improved build scripts to optionally handle registries.
Scripts also now tag images with timestamp as the tag for easier upgrading.

Updated build-all-and-refresh-pods-local.sh to install/upgrade in one go.

Co-authored-by: xiani <xianionline@gmail.com>