Fixed errors in WebRTC connections. (#4851)
Added a list of public STUN servers to engine networking constants.
createTransport() in SocketWebRTCClientFunctions.ts adds this as 'iceServers' to transportOptions
when creating transports, so that the RTCPeerConnections that are created have specific STUN servers
to get their public IP from. (prior to this there were no explicitly defined STUN servers, and roughly
half the time WebRTC connections failed because they weren't getting a public IP for the client).

Updated get-local-server-ip.ts to use internalIp.v4() if 'localhost' is specified as GAMESERVER_HOST.
If any other value is used, then it just uses that value.