Removed server/gameserver/client/analytics enabled config values.
I think this was a relic from when gameserver and api server were the same codebase.
It was functionally doing nothing other than messing up deployments when GAMESERVER_ENABLED
or SERVER_ENABLED wasn't set properly, which would just cause either to not start properly and
result in endless loops of failures when liveness probes would fail. This got copied to client
and analytics without doing anything useful there, either.

If the respective pods are running, it can be assumed that the service is enabled, no additional
configuration is necessary.