Added user-api-key service and refactored authentication. (#4689)
Created user-api-key service. This is mostly just a table containing a token that's
linked to a user. API keys are created when a non-guest user is created, or when they
log in for the first time. External requests to user_api_key.PATCH, unless called by an
admin with an ID, will update the token on the user_api_key for the authenticating user.

Updated authentication flow to incorporate API keys. API keys are sent via header
'Authorization: Bearer <token>', just like JWTs. New authenticate.ts first tries to validate
against the user-api-key table, and if it matches, returns the user that owns that token.
If there is no match, it then tries to do JWT authentication.

Whether authenticating with an API key or a JWT, instead of just passing the identity-provider onto
context.params, authenticate.ts now also fetches the authenticated user and puts that at context.params.user,
so that services don't have to do the lookup themselves.

Modified extractLoggedInUserFromParams to return the user from the params instead of the identity-provider. Changed
all implementations of this so that references to loggedInUser.userId become loggedInUser.id, no additional fetches of
the user are performed, etc.

Replaced all other calls to authenticate('jwt') inside service's hooks.ts with authenticate(), referring to
authenticate.ts.

The user profile menu now has a button very similar to the 'show user ID' one that shows the user's API key, and allows
for it to be easily copied to the clipboard. There's also a button to refresh the key.