Forward pose data immediately to other clients
The security model needs to be improved (we should encode user ids so the consumer knows which user sent this data), but for now this results in much lower latency than waiting for the next frame to process in the gameserver. This also means the server has to do much less work, as it is only responsible for encoding the network objects it has authority over (rather than continually decoding and re-encoding everything).