Fixed bug in producer creation/resumption.
Client-side loop to wait for audio/video producers to start was trying to create
producers on every iteration. This resulted multiple producers being made on the
backend, with the last one being set as the client's producers. When the client
sent signals to resume producers, they had the ID of the first one to be created,
leading to the gameserver not being able to find a matching producer and crashing.

Loop now only creates producer on first iteration. gameserver now has a check that
a hostClient was found before trying to send a message over it.