Switched Helm references from `mariadb` to `sql`.
Using `mariadb` for all SQL database configuration in the Helm chart suggested that only MariaDB
was supported, which is not the case. The charts have been changed to use `sql.<value>`, and
deprecated values have been removed, as well as the tree flattened and some value names made shorter.

The chart still has support for `mariadb` so that existing deployments don't error out, but `sql` is
top preference.

Removed some old k8s files that were extremely outdated and hadn't been used in years.

Added nil capture groups around all checks for services existing in Helm files, so that if the service
isn't defined at all, it resolves to false and doesn't error out from trying to read sub-values of
a nonexistent object.