Updated instructions on running stack entirely in local Docker.
Tested how to run entire stack in Docker containers and documented the process.
Almost everything can be automated, but the user needs to get and supply their
local IP address. When Docker runs on the host network, internalIp.v4() returns
undefined.

Added a mode 'serve-local' to start-server.sh, so that local files can be served by
a Docker container.

Expanded root Dockerfile to accept/use more args. They are needed for building the image
for local Docker usage.

Added a run of 'check-db-exists' to root Dockerfile. This will seed the database with certain
necessary values for building the clien (though a later run of the API server with
FORCE_DB_REFRESH=true is still required to seed everything), and also uploads the default static
files to server/upload in the image, so that the serve-local container can serve them.