Fixed bug in make-user-admin.js (and other similar scripts).
Some of the make-user-admin scripts were using dotenv instead of dotenv-flow, and were
ignoring .env.local files. In cases like WSL where the DB address is not 127.0.0.1, the
script was not using the configured address from the proper .env file.

Made make-user-admin.js check for existing scopes so that they are not duplicated, and
print a different message if the user does not exist.