Fixed some bugs with authentication settings. (#5418)
Discord key/secret in client page were mistakenly changing Facebook values.

Improved patching of authentication settings to handle enabling of not-previously-enabled
providers.  One, providers with non-settable oauth fields like `scope` were not having that
inserted into the oauth settings; two, callback URLs for newly-enabled providers were not
being set in the callback settings. authentication-setting patch() now iterates over all the
providers in the patch data and adds scope/custom_data to the oauth entry if present in the
the config for that provider, and adds a callback URL for it from the config as well.

Co-authored-by: Zulqarnain Hanif <30355034+zulqarnainhanif@users.noreply.github.com>