Fixed a bug with seeding values.
Seeding didn't take into account uniqueness. e.g., if a location seed shared a name with
a location that existed under a different ID, the seed ID lookup would show no matches and
try to insert a new location, and would fail on the uniqueness constraint.

ID lookup has been removed for seeding. Now, if it's not a settings service, it tries to find
a unique field on the model and just find a match on that. If there's no unique field, then it
does what it was doing before, and just searches on all the field values specified in the seed.