Fixed a bug with minikube client local file loading.
getCachedAsset.ts always uses cacheDomain for generating URLs. On minikube, though, when using local storage,
the client and server need different addresses. The server needs to use 10.0.2.2 to punch out of minikube to
the host machine's localhost, while the client just needs to use localhost.

New logic assumes that localStorageProvider will be 10.0.2.2, and so alterations to that are done if new parameter
'clientFetch' is explicitly true (defaults to false). scene.class.ts now accepts query parameter clientFetch and passes
it on.

Fixed a bug in adminLocations fetching that rebrands location_setting to locationSettings, as we do with non-admin locations.