Discord bot integration (#4981)
* Added DiscordBotAuth service

* Made discord-bot-auth work.

Service tries to authenticate bot with token against Discord API.
If it's not a bot, returns an error. If it is a bot, searches for
an XREngine identity-provider with that Discord ID. Returns that if
it exists, makes it if not.

Updated identity-provider.class.ts to handle bot user creation. Added
'bot' to user-role.seed.ts

* fix errors

* format

* tsc fixes

Co-authored-by: Josh Field <joshfield999@gmail.com>