Fix bug when trying to delete broken project
Projects are ending up in a broken state, without an xrengine.config.ts file, but cannot be deleted.
getProjectConfig() returns null if it can't find that file, but the project.remove handler was expecting
it to be not null. Adding a conditional before the references to projectConfig.onEvent should fix this
and allow the project record to be deleted.