Renamed gameserver -> instanceserver
Since the functionality of 'gameserver' wasn't always to enable game behavior,
the decision was made to rename it 'nameserver'.

Breaking changes:
This includes the tables 'gameServerSetting' and 'gameserver_subdomain_provision'.
The latter shouldn't be a problem, since the table is automatically made and nothing
should be interacting with it (it's legacy behavior that was kept around just in case).
The former will present a slight problem - the entry in gameServerSetting should be
copied verbatim to the new table 'instanceServerSetting'.