Added hack fix for nested paths in importSystem. (#6239)
rollup's dynamic-import-vars plugin cannot replace a variable in its path with anything more than a single
level deep '*'. The prior implementation of importSystem was using a potentially multi-level path, leading
to client-side behavior of not finding any systems that weren't in the root of the project.

Added a hack fix of handling files up to five levels deep by using a different import string based on splitting
the path on '/' and manually inserting a '/' between them in the import statement. There didn't seem to be a way
to design this to work with any depth, but five seems like more than enough and there's an error returned if
someone tries to do more.