Added project permissions and GitHub push integration.
Project permissions can now be made to authorize a non-admin access to a project.
Project owners and users can access a project and its files in the editor. Owners
can give users new permissions, edit the owner/user status of existing users, and
remove users' permissions. Users can just access the project.

Users' GitHub OAuth token is now saved when they log in. When projects are being
fetched, it checks which repos that the GH app has been installed in they have
write access to. Those that match are considered permitted at the user level.

Users with permissions on a project can push that project to GitHub. This will
attempt to use their personal GH token to make the new commit and push, but if
they do not have GH credentials, then it will push it via the app's credentials.

Users with project-permissions can update the repositoryPath of the project. This
must be to a repo that the user has access to.