Fixed a couple of bugs related to chat.
channel.find() was not using the sequelize include statements, and was not
returning the messages (nor other populated information) on the response.
This was causing bugs on the frontend, which expects the messages to be part
of the response.

InstanceChat was only rendering the 'open world chat' button if `unreadMessages`
were true. This seems like undesired behavior - a user should be able to open
the world chat even if there aren't any new messages.