Fixed bugs in instance channel fetching and display. (#6394)
Sometimes, the client recognizes a connection to an instance server has been made
before the initial 'joined' message has been created, and so when the instance channel
is fetched, there are no messages. Added a check for number of messages that triggers
a refetch in 500 ms until there are messages to display.

Added some checks to InstanceChat/index.tsx that messages[index -1] exists before trying
to call isLeftOrJoinText on that; the above bug brought to light a bug here where that
call would throw an error since there was no prior message to call on.