Fixed a bug with OAuth login.
loginUserByJWT had a race condition where it hadn't always finished setting the new JWT in localStorage before
redirecting the user back to the page they started from. In this case, doLoginAuto would use their old JWT, and
which, if they had logged in from being a guest, would not match to the identity-provider/user they intended to
log in as. If the identity-provider already existed, then the identity-provider/guest user the old JWT was
generated for would no longer exist, and they'd be logged in as a new guest.

loginUserByJWT now uses setInterval to wait until the token stored in localStorage matches the one that it
intends to set there, and only redirects when they match, i.e. localStorage has been fully updated.