Added project permissions and GitHub push integration (#6380)
* Added project permissions and GitHub push integration.

Project permissions can now be made to authorize a non-admin access to a project.
Project owners and users can access a project and its files in the editor. Owners
can give users new permissions, edit the owner/user status of existing users, and
remove users' permissions. Users can just access the project.

Users' GitHub OAuth token is now saved when they log in. When projects are being
fetched, it checks which repos that the GH app has been installed in they have
write access to. Those that match are considered permitted at the user level.

Users with permissions on a project can push that project to GitHub. This will
attempt to use their personal GH token to make the new commit and push, but if
they do not have GH credentials, then it will push it via the app's credentials.

Users with project-permissions can update the repositoryPath of the project. This
must be to a repo that the user has access to.

* fix project page crash

* Moved Project feathers listener into useApiListeners, added calls to it.

Renamed props on new Dialogs to match recent pattern change.

* Renamed modal to drawer, unified set and update github link into single drawer, misc changes for add project drawer

* Changed column name for repo link

* Allow admins to create project permissions even if there are none on a project.

Fixed population of hasWriteAccess to project.

If no permissions on a project when one is being added, that person is automatically an owner.

* Misc fixes

* Fixed icon color

* Repositioned cancel button to make ui consistent

* Updated styles to match admin changes

* Fixed bugs.

Permissions modal in admin page had some issues with state of active project pointing to another
project after update. Now tracking projectId separately and using that to ensure correct project
is active.

Fixed bugs with non-installed project showing clickables that they should not.

Deleted files were not being deleted in GH tree. They now are.

Fixed bugs with project-permissions not cascade-deleting on project deletion.

* fixup

* Fixed format issue

* Fixed icon button color

* Fixed build errors

* Fixed build error

* Fixed more errors

Co-authored-by: HexaField <joshfield999@gmail.com>
Co-authored-by: Hanzla Mateen <hanzlamateen@live.com>