Fixed multiple login/logout/account deletion issues. (#6530)
A bug was noticed in production where identity-providers were lacking userIds. This appears to have been caused by
the relationship in that database being set to onDelete: setNull instead of onDelete: cascade. This has been fixed
in that DB schema, and onDelete: cascade has been added to identity-provider association to reinforce it.

Added some bug-handling to oauth provider code to deal with this should it arise again. If the existing identity-provider
is missing a userId, it makes a new user and associates it to the identity-provider before continuing.

Users were confused by being able to remove all of their oauth connections and thus their account. The client will no
longer let the user remove their last OAuth connection if it's their last identity-provider, and notifies the user of this.

Added a deletion flow to the user ProfileMenu.

Co-authored-by: Gheric Speiginer <gheric.speiginer@gmail.com>