Fixed some bugs in admin project page related to selected project state.
Admin project table was referencing project state in isolated handler functions.
If a function is declared, it can't reference a useState() state directly, since the value
gets set when the function is declared and won't change when the state changes. Changed
to add a Ref to the state that the functions can use, as <ref>.current does change.

Restored a useEffect to update selected project state if its store value changes.