Made location share modal work with invite changes.
Without spectate on, location share modal will generate link with user's inviteCode.
With spectate on, it will generate a link without an inviteCode but with spectate=userId.

Invite is now an 'instance' invite. Added spawnType: 'spectate', spawnDetails for it are
{ spectate: <userId> }. Updated invite flow to handle spectate invites.

Fixed issue with new way of handling invite finds. It was incorrectly restricted to admins,
but it's needed for checking if an invite already exists. Added a new query param 'existenceCheck'
that will leave the userId on the query if true, so that admins can get all invites in the admin page
but regular users can still check if an invite exists.