Invite system upgrades (#6575)
* Expanded invite system, added location authorized users.

Consolidated admin invite page to a single table. Having separate tables for sent and received
does not make sense for admins - they just see all invites.

Added admin invite service for retrieving all invites and patching them. Admin invite page now
allows for creating new-user, location, and instance invites, which are all new inviteTypes.
new-user just makes a user if they don't exist and can optionally make them an admin immediately.
Location and instance invites have a dropdown to select the location/instance. They can optionally
allow for the user being spawned at another user's location (via that user's invite code) or at a
selected spawn point. Admin invite page also allows for patching a single invite, and for bulk
deleting invites. Invites can now be set to not be deleted on use.

Fixed some backend bugs with teleporting in the process of adding spawning invites. Also added
function to find spawn point by ID and teleport to that.

Doing a FIND on invites without specifying sent or received has new logic that is only available
to admins.

Added location-authorized-user service for private locations. If a user is not authorized to
be at a location, the client will not render it nor attempt to connect to any instanceservers.
Currently, the only way to become an authorized user is to get an invite to that location
(or an instance of it).

* Made location share modal work with invite changes.

Without spectate on, location share modal will generate link with user's inviteCode.
With spectate on, it will generate a link without an inviteCode but with spectate=userId.

Invite is now an 'instance' invite. Added spawnType: 'spectate', spawnDetails for it are
{ spectate: <userId> }. Updated invite flow to handle spectate invites.

Fixed issue with new way of handling invite finds. It was incorrectly restricted to admins,
but it's needed for checking if an invite already exists. Added a new query param 'existenceCheck'
that will leave the userId on the query if true, so that admins can get all invites in the admin page
but regular users can still check if an invite exists.

Co-authored-by: Kyle Baran <kbaran@hyperconstruct.io>