Added timing to invites.
Invites can now be set to start and/or end at specific times. If a user tries to accept them before
the start time or after the end time, it will not go through and they'll be informed of that fact.

Consolidated some duplicated code in accept-invite.class.ts and invite create/update modals.