Fixed a bug with restarting user video. (#6677)
useEffect watching videoStream.track.id in UserMediaWindow is not firing when a new track
replaces a paused producer's ended video track. For now, made pauseProducer in instanceserver
be emitted to all clients so that self user's handler can act on that to pause videoStream,
which will make resuming it trigger needed updates.