Fixed a bug in UserMediaWindow with pausing producers. (#6693)
Pausing self producer was causing all consumers to be paused. pauseProducer listener
needed to check if consumers' producerId matched the ID of the producer being paused.

Added color to paused/active icons in UserMediaWindow. Gray-on-gray was hard to make out.

Attempted to fix a bug where client gets a bad instance provision. ID was getting updated
in URL when user's instanceId was patched, but Engine.instance.currentWorld._worldHostId
was staying on the invalid one. This led to infinite failed attempts to get the channel
since the world instanceId was incorrect. Made patchUser handler update currentWorld's
hostIds when this changes.