Fixed bugs in instance connections. (#6709)
If a user disconnected right after connecting to a fresh instanceserver, the disconnect logic would
not work properly. Since app.instance had not been set, it would not call handleUserDisconnect.
Added a one-second timeout to wait for possible server startup to resolve with app.instance populated.

Updated check to call shutdownServer on peers.size <= 1. On near-simultaneous connection/disconnection,
the server peer may not have been created, leaving the number of peers 0. Prior check was if it was
exactly 1, which would not be true.

Fixed a bug in instance-provision where calling get(channelId) or get(locationId) wasn't being
try/caught, so if there was no channel, it was throwing an uncaught error.