Fixed a bug with instance provisioning and unreachable instanceservers. (#6720)
* Fixed a bug with instance provisioning and unreachable instanceservers.

On a couple of occasions, instanceservers were not externally reachable -
they could be kubectl exec'd into and responded to internal requests, but
any request from outside was hanging. Updated instance-provision to make
a request to the server it's about to return. If there's an error, or it
does not receive any response within two seconds, it removes the instance
record, deletes the pod in a K8s environment, and then tries to get another
free instanceserver.

* Made unreachable timeout a configurable serverSetting