Converted userRole: admin to an admin scope. (#6727)
* Converted userRole: admin to an admin scope.

Admin status was defined by user.userRole being 'admin'. This has been changed so that
admin status is a scope 'admin:admin'.

Made a script to convert all existing `userRole: admin`s to `userRole: user` and add an
admin scope for them.

* Replaced userRole with isGuest.

Instead of userRole table related to user table, with potential for several roles,
now the only thing tracked on the user is whether they're a guest.

Changed convert-admin-to-scope.js to convert-user-roles.js, which will make guests
to isGuest: true and non-guests to isGuest: false.