Converted avatar logic to use a new avatar table.
Maintaining user avatars just as a pair of static-resources referenced by name
was cumbersome. Added a new table `avatar` that has a modelResourceId and
thumbnailResourceId, which point to the associated static-resource. User's
avatarId is now the ID of the avatar row.

Avatar name is no longer unique. identifierName is formed from concatenating
the avatar name and ID, and is used for naming the associated static resource
files, e.g. avatar1_<UUID>.

Static resource uploading is now performed separately from the avatar creation
handler. The avatar row has to be created so that the identifierName can be
generated, then the assets can be uploaded to the storage provider using
identifierName, then finally the avatar is patched with the static resource IDs.