Fixed issues with self video element and audio consumer pausing (#6738)
When a client is alone in a media channel, trying to pause and then resume cam video
was resulting in a blank video for themself. The useEffect listening for videoStream.track.id
changing is, for some reason, not triggering, despite re-render logging showing that the value
does change. If there's at least one other UserMediaWindow, though, it resumes normally.

This might be due to tracking an object that isn't fully in React state (though the producer
is set in a state object). Added new state to explicitly track video/audio track ID, and made
useEffects dependent on that, which seems to work better.

Added a state object to track when a video is ready for viewing to avoid flickering when
video is resumed. Before, the ended track was briefly seen before being replaced; now nothing
is shown until the new track has replaced it.

Fixed bug where a consumer pause could be overridden by the producer pausing and resuming.
This was due to the instanceserver emitting consumer pauses/resumes on producer pause/resume.
Deleted those since all connected clients get producer pauses/resumes and can make their own
decisions about whether to pause/resume consumers.

Co-authored-by: Gheric Speiginer <gheric.speiginer@gmail.com>