Fixed some issues with party and scope tests, and a bug that was caught from them.
Scope tests had intermittent failures that seem to be from not re-fetching the
user after adding scopes to them.

Party delete test had intermittent failures due to a race-condition bug in the logic.
When the party deletion went to delete the users, deleting the party owner would trigger
another user to be made party owner, but they may have been deleted before that patch
triggered. Changed the parameter 'skipPartyDelete' that's passed to party-user.delete
to 'deletingParty', and made transfer of party ownership skipped if this parameter is true.