Updated project GitHub integration to use branches.
Having every installation of a project push to the main branch in GitHub could be problematic.
If different deployments do different things with the base project, pushing those changes to
the main Github branch could overwrite what someone else had pushed up. To counter this, and
make it far less likely that the main branch is inadvertently overwritten, each deployment
now pushes to a branch named after its own releaseName, i.e. <releaseName>-deployment.

The 'update' functionality now pulls from that deployment branch. A new feature was added
to reset the deplyoment branch to the main branch; this clones the repo and force-pushes
the main branch to the deployment branch.